app-macos/requirements.txt
rumps>=0.4.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
requests>=2.28.0
py2app>=0.28.0
